Refactor containers #49
